Separate Rust and Python network modules
